+++ src/main/java/lgk/nsbc/ru/App.java
@@ -5,20 +5,20 @@ import com.vaadin.annotations.VaadinServletConfiguration;
 import com.vaadin.server.VaadinRequest;
 import com.vaadin.server.VaadinServlet;
 import com.vaadin.ui.Button;
-import com.vaadin.ui.ComboBox;
 import com.vaadin.ui.TabSheet;
 import com.vaadin.ui.UI;
 import lgk.nsbc.ru.backend.ConsultationManager;
 import lgk.nsbc.ru.model.ConsultationModel;
+import lgk.nsbc.ru.presenter.CalendarPresenterImpl;
 import lgk.nsbc.ru.presenter.ConsultationPresenter;
-import lgk.nsbc.ru.view.CalendarView;
+import lgk.nsbc.ru.view.CalendarViewImpl;
+import lgk.nsbc.ru.view.CalendarViewOld;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.ArrayList;
 
 /**
  *
@@ -40,18 +40,23 @@ public class App extends UI {
 		SessionManager sessionManager = new SessionManager();
 		PatientsPresenter patientsPresenter = new PatientsPresenter(model, patientsManager, sessionManager, lgkSessionId);
 		patientsPresenter.start();*/
-		ConsultationModel consultationModel = new ConsultationModel();
-		CalendarView calendarView = new CalendarView(consultationModel);
+		/*ConsultationModel consultationModel = new ConsultationModel();
+		CalendarViewOld calendarView = new CalendarViewOld(consultationModel);
 
 		ConsultationManager consultationManager = new ConsultationManager();
 		ConsultationPresenter consultationPresenter = new ConsultationPresenter(consultationModel,consultationManager);
 		calendarView.setPresenter(consultationPresenter);
-		consultationPresenter.start();
+		consultationPresenter.start();*/
+
+		ConsultationModel consultationModel = new ConsultationModel();
+		ConsultationManager consultationManager = new ConsultationManager();
+		CalendarPresenterImpl calendarPresenter = new CalendarPresenterImpl(consultationModel,consultationManager);
+		calendarPresenter.start();
 
 		//setContent(calendarView);
 		TabSheet tabSheet = new TabSheet();
 		tabSheet.setHeightUndefined();
-		tabSheet.addTab(calendarView,"Радиохирургия");
+		tabSheet.addTab((CalendarViewImpl)calendarPresenter.getCalendarView(),"Радиохирургия");
 		tabSheet.addTab(new Button("кнопка"),"Очные");
 		tabSheet.addTab(new Button("кнопка"),"Заочные");
 		tabSheet.addTab(new Button("кнопка"),"Онкология");
