+++ src/main/java/lgk/nsbc/ru/view/AbstractView.java
@@ -2,10 +2,10 @@ package lgk.nsbc.ru.view;
 
 import com.vaadin.ui.CustomComponent;
 
-public abstract class AbstructView<Model> extends CustomComponent {
+public abstract class AbstractView<Model> extends CustomComponent {
 	protected final Model model;
 
-	public AbstructView(Model model) {
+	public AbstractView(Model model) {
 		this.model = model;
 	}
 }
